Abstract utilities
